Forwards the :database option to the migration collection calls
This is a way we can support dynamic multitenancy. If you pass in the :database option into the `migrate/1` function, it will forward it onto each migrations `up` function, but we also need it to use the database option for each of the calls to the migration collection, so that each tenant's migration states are kept separate.