Let `Mongo.Repo` errors flow like with `Mongo` functions (#219)
When `Mongo` returns a `Mongo.Error`, we get match errors inside `Enum`,
which are hard to read and sometimes misleading. This patch changes
signatures to let the errors flow. This is argably a breaking change,
but errors will happen anyway in problematic scenario.

As alternative, `Mongo.Error` could be `raise`d and signtures are kept
the same. At least errors will be more clear to library users.